celery_app/celery_requirements.txt
celery==5.3.1
httpx==0.24.1
python-dotenv==1.0.0
